macOS  - dmg 'installer'

+++ scripts/release_macOS.sh
@@ -13,7 +13,10 @@ BUILD="$ROOT"/build
 TARGET=rclone-browser-$VERSION-macOS
 APP="$TARGET"/"Rclone Browser.app"
 
-rm -rf "$BUILD"
+if [ -d "$BUILD" ]; then
+  rm -rf "$BUILD"
+fi
+
 mkdir -p "$BUILD"
 cd "$BUILD"
 # brew install cmake qt5
@@ -27,7 +30,9 @@ cd ../..
 
 mkdir -p release
 cd release
-rm -rf "$TARGET"*
+if [ -d "$TARGET" ]; then
+  rm -rf "$TARGET"*
+fi
 mkdir "$TARGET"
 cp "$ROOT"/README.md "$TARGET"/Readme.txt
 cp "$ROOT"/CHANGELOG.md "$TARGET"/Changelog.txt
@@ -47,3 +52,14 @@ EOF
 7za a -mx=9 -r -tzip "$TARGET".zip "$TARGET"
 
 ## gpg --detach-sign "$TARGET".zip.sig "$TARGET".zip
+
+# prepare dmg 'installer'
+# https://github.com/LinusU/node-appdmg
+if [ -d "Rclone Browser.app" ]; then
+  rm -rf "Rclone Browser.app"
+fi
+cp -r "$TARGET"/"Rclone Browser.app" .
+cd ../scripts
+appdmg ../appdmg.json "$TARGET".dmg
+cd ../release
+rm -rf "Rclone Browser.app"
